sections of details component are separated as subcomponents
